# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  models.ZenviaMessage:
    properties:
      externalId:
        type: string
      messageId:
        type: string
      to:
        type: string
    type: object
  models.ZenviaMessageStatus:
    properties:
      causes:
        items:
          $ref: '#/definitions/models.ZenviaStatusCause'
        type: array
      code:
        type: string
      description:
        type: string
    type: object
  models.ZenviaStatusCause:
    properties:
      details:
        type: string
      reason:
        type: string
    type: object
  models.ZenviaWebhookRequest:
    properties:
      message:
        $ref: '#/definitions/models.ZenviaMessage'
      messageStatus:
        $ref: '#/definitions/models.ZenviaMessageStatus'
      type:
        type: string
    type: object
  models.ZenviaWebhookResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: Basic authentication with client_id as username and client_secret as
    password (for token generation only).
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WSICRMREST API
  version: 1.26.4.27
paths:
...
  /webhook/zenvia/email:
    post:
      consumes:
      - application/json
      description: |-
        Processes email status events received from Zenvia webhook. Updates email status, logs API history, and creates occurrences for bounced emails.

        **Behavior for Unknown Message IDs:**
        - Returns HTTP 200 with success=true (prevents Zenvia retries)
        - Logs warning with messageId and error details
        - Stores webhook request in WSREQUISICOES table for audit
        - Message: "Mensagem não encontrada no banco de dados"

        **Supported Events:**
        - sent (121 - Agendado)
        - delivered (122 - Entregue)
        - read/clicked (123 - Aberto)
        - rejected/not_delivered (124 - Não Entregue)

        **Security:** No authentication required (public webhook endpoint)
      parameters:
      - description: Zenvia webhook payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.ZenviaWebhookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook processed successfully (includes cases where message
            ID is not found)
          schema:
            $ref: '#/definitions/models.ZenviaWebhookResponse'
        "400":
          description: Invalid request body or JSON
          schema:
            $ref: '#/definitions/models.ZenviaWebhookResponse'
      summary: Zenvia Email Webhook
      tags:
      - Webhooks
...
schemes:
- http
- https
securityDefinitions:
  BasicAuth:
    type: basic
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...